layout/generic/test/test_bug735641.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=735641
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 735641</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=735641">Mozilla Bug 735641</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 735641 **/

SimpleTest.waitForExplicitFinish();

var w;

function runTest() {
  try {
    var s = w.document.getSelection();
    s.selectAllChildren(w.document.body);
    synthesizeMouse(w.document.body,1,1,{},w);
    ok(s.isCollapsed,"mouse click on <body> did reset the selection in image document");
  } finally {
    w.close();
  }
  SimpleTest.finish();
}

w=window.open("file_Dolske.png","_blank");
w.addEventListener("load", function() {setTimeout(runTest,0)}, false);

</script>
</pre>
</body>
</html>
